refactor(vertex-forager): split runtime domain models
